move geo files out of geo submodule
