--- rosys/geometry/geo/__init__.py
@@ -1,11 +1,0 @@
-from .fixpoint import Fixpoint
-from .geo_point import GeoPoint
-from .geo_pose import GeoPose
-from .geo_reference import GeoReference
-
-__all__ = [
-    'Fixpoint',
-    'GeoPoint',
-    'GeoPose',
-    'GeoReference',
-]
